move fixtures to proper directory
