+++ src/lexer/source-token/identifier/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "bun:test";
-import { fakePos1, fakePos2, fakeRange } from "../index.test.fixture";
+import { fakePos1, fakePos2, fakeRange } from "../../../util/position/index.test.fixture";
 import type { IdentifierToken, KeywordToken } from "./";
 import { createIdentifierToken, createKeywordToken } from "./";
